An author's note was added, with my peculiar way of making a random creator's note, haha.
I added a new, handwritten creator's note to the README. It's just fun information about the project, the "team," and a bit about my life as the creator of the optimized engine. Basically, I want the README to be more personal, fun, and relatable, and for kids to feel like they're reading a chat message.